+++ test/ct/wait-for.sh
@@ -1,6 +1,8 @@
+#!/bin/bash
+
 # MIT License
 #
-# (C) Copyright [2021] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2022] Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -20,34 +22,27 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 
-# Dockerfile for running bss unit tests.
+# wait-for.sh; used by runCT.sh to make sure HSM has been populated with data before running.
+echo "Initiating..."
+URL="http://cray-smd:27779/hsm/v2/State/Components"
+sentry=1
+limit=200
+while :; do
+  length=$(curl --silent ${URL} | jq '.Components | length')
 
-### build-base stage ###
-# Build base just has the packages installed we need.
-FROM arti.dev.cray.com/baseos-docker-master-local/golang:1.16-alpine3.13 AS build-base
+  if [ ! -z "$length" ] && [ "$length" -gt "0" ]; then
+    echo $URL" is available"
+    break
+  fi
 
-RUN set -ex \
-    && apk -U upgrade \
-    && apk add build-base
+  if [ "$sentry" -gt "$limit" ]; then
+    echo "Failed to connect for $limit, exiting"
+    exit 1
+  fi
 
-### base stage ###
-# Base copies in the files we need to test/build.
-FROM build-base AS base
+  ((sentry++))
 
-RUN go env -w GO111MODULE=auto
+  echo $URL" is unavailable - sleeping"
+  sleep 1
 
-# Copy all the necessary files to the image.
-COPY cmd $GOPATH/src/github.com/Cray-HPE/hms-bss/cmd
-COPY pkg $GOPATH/src/github.com/Cray-HPE/hms-bss/pkg
-COPY vendor $GOPATH/src/github.com/Cray-HPE/hms-bss/vendor
-COPY .version $GOPATH/src/github.com/Cray-HPE/hms-bss/.version
-
-### testing stage ###
-FROM base AS testing
-
-WORKDIR /go
-
-# Run unit tests.
-RUN set -ex \
-    && go clean -testcache \
-    && go test -cover -v github.com/Cray-HPE/hms-bss/cmd/boot-script-service
+done