+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.16.0] - 2022-03-03
+
+### Changed
+
+- converted image builds to be via github actions, updated the image links to be in artifactory.algol60.net
+- added a runCT.sh script that can run the tavern tests and smoke tests in a docker-compose environment
+
 ## [1.15.0] - 2022-01-07
 
 ### Changed